trendstack/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trendstack"
version = "0.1.0"
description = "Algorithmic Trading System with TrendStack Architecture"
authors = [{name = "TrendStack Team", email = "dev@trendstack.com"}]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # Core trading libraries
    "backtrader>=1.9.78",
    "pandas>=2.3.0",
    "numpy>=2.3.0",
    "scipy>=1.16.0",
    "yfinance>=0.2.65",
    
    # Technical analysis
    "ta>=0.11.0",
    
    # Visualization
    "matplotlib>=3.10.0",
    "plotly>=6.2.0",
    "seaborn>=0.13.0",
    "quantstats>=0.0.68",
    
    # Data processing
    "pydantic>=2.11.0",
    "pyyaml>=6.0.0",
    "requests>=2.32.0",
    
    # Utilities
    "loguru>=0.7.0",
    "schedule>=1.2.0",
    "prefect>=3.4.0",
    
    # Communication
    "python-telegram-bot>=13.15,<14.0",
    
    # Trading platforms
    "MetaTrader5>=5.0.5120",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

research = [
    "jupyter>=1.0.0",
    "notebook>=6.5.0",
    "ipywidgets>=8.0.0",
]

monitoring = [
    "prometheus-client>=0.22.0",
    "grafana-api>=1.0.3",
]

[project.urls]
Homepage = "https://github.com/shaxb/supurgi-v5"
Repository = "https://github.com/shaxb/supurgi-v5.git"
Issues = "https://github.com/shaxb/supurgi-v5/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["trendstack*"]

[tool.black]
line-length = 88
target-version = ['py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=trendstack --cov-report=term-missing"
testpaths = [
    "cores/*/tests",
]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
